+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file only records notable changes. Not synchronized with all releases and t
 - main - not yet released
 - v3.8.0 - not yet released
   - Deprecate some options to prepare v4: [#1045](https://github.com/kachick/wait-other-jobs/pull/1045)
+  - Add `outputs.parameters`
 - v3.7.0
   - Begins outputting job summary: [#1026](https://github.com/kachick/wait-other-jobs/pull/1026)
 - v3.6.0